% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgverse.R
\name{pkgverse}
\alias{pkgverse}
\title{pkgverse}
\usage{
pkgverse(pkg, pkgs)
}
\arguments{
\item{pkg}{Name of your set of packages. It's recommended that users append
this name with something like 'verse' or otherwise provide some kind of
explicit/obvious indicator that the package name is a stand in for a
selection of packages.}

\item{pkgs}{Character vector of package names.}
}
\value{
Installs a package of desired name, which, when loaded, will load,
  handle, and display conflicts of the packages supplied via \code{pkgs}.
}
\description{
Create your own pkgverse
}
\examples{
\dontrun{

## example: web scraping package universe
pkgverse("webscraperverse", c("xml2", "rvest", "httr", "RSelenium"))

}
}
