% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgverse.R
\name{pkgverse}
\alias{pkgverse}
\title{pkgverse}
\usage{
pkgverse(pkg, pkgs)
}
\arguments{
\item{pkg}{Name of your -verse (verse will be appending to whatever you enter here).}

\item{pkgs}{Character vector of packages.}
}
\value{
Installs a package of desired name (with 'verse' attached to what is supplied)
}
\description{
Create your own pkgverse
}
